+++ deptran/rcc_service.h
@@ -202,7 +202,7 @@ class ClassicServiceImpl: public ClassicService {
 
   void OVCreateTs(const cmdid_t &txn_id,
                  rrr::i64 *timestamp,
-                 rrr::i64 *server_id,
+                 rrr::i16 *server_id,
                  rrr::DeferredReply *defer) override ;
